+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/MessageListView.kt
@@ -288,7 +288,8 @@ public class MessageListView : ConstraintLayout {
                             hasTextToCopy = message.text.isNotBlank(),
                             suppressThreads = adapter.isThread || message.isInThread(),
                         ),
-                        requireStyle()
+                        requireStyle(),
+                        messageListItemViewHolderFactory
                     )
                     .apply {
                         setReactionClickHandler { message, reactionType ->
@@ -388,7 +389,8 @@ public class MessageListView : ConstraintLayout {
                         hasTextToCopy = false, // No effect when displaying reactions
                         suppressThreads = false, // No effect when displaying reactions
                     ),
-                    requireStyle()
+                    requireStyle(),
+                    messageListItemViewHolderFactory
                 ).apply {
                     setReactionClickHandler { message, reactionType ->
                         messageReactionHandler.onMessageReaction(message, reactionType)
@@ -608,8 +610,8 @@ public class MessageListView : ConstraintLayout {
         initAdapter()
 
         messageListViewStyle = requireStyle().copy(
-            replyEnabled = requireStyle().replyEnabled && channel.config.isRepliesEnabled,
-            threadsEnabled = requireStyle().threadsEnabled && channel.config.isRepliesEnabled,
+            replyEnabled = requireStyle().replyEnabled,
+            threadsEnabled = requireStyle().threadsEnabled && channel.config.isThreadEnabled,
         )
     }